pyproject.toml
[project]
name = "resume-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pdf2image==1.17.0",
    "pdfplumber==0.11.4",
    "pytesseract==0.3.13",
    "python-dotenv==1.0.1",
    "streamlit==1.32.0",
]
